isThisElementOnScreen, scrollCheck, getPageOffsetGenerator, getPageOffset -> lazyload#...
