+++ test/unit/core.js
@@ -210,50 +210,4 @@ describe('Imager.js', function () {
       expect(new Imager(imgrOptions)).to.have.property('devicePixelRatio', 2);
     });
   });
-
-  describe('isThisElementOnScreen', function () {
-    var offsetStub, imgr;
-
-    beforeEach(function () {
-      imgr = new Imager('.delayed-image-load');
-      offsetStub = sandbox.stub(imgr, 'viewportHeight', 1024);
-
-
-    });
-
-    it('should not detect an element if itself and its parent outreach the viewport height', function () {
-      var element = {
-        offsetTop:    100,
-        offsetParent: {
-          offsetTop: 1024
-        }
-      };
-
-      expect(imgr.isThisElementOnScreen(element)).to.equal(false);
-    });
-
-    it('should detect an element on screen if its offsetPosition is located within the viewport', function () {
-      var element = {
-        offsetTop:    100,
-        offsetParent: {
-          offsetTop: 0
-        }
-      };
-
-      expect(imgr.isThisElementOnScreen(element)).to.equal(true);
-    });
-
-    it('should detect an element on screen if its offsetPosition is out of the viewport but within the viewport+lazyloadOffset', function () {
-      var element = {
-        offsetTop:    100,
-        offsetParent: {
-          offsetTop: 1024
-        }
-      };
-
-      imgr.lazyloadOffset = 300;
-
-      expect(imgr.isThisElementOnScreen(element)).to.equal(true);
-    });
-  });
 });
